+++ packages/kitchensink/service-fastify-cors/feature-custom-routes.js
@@ -1,0 +1,11 @@
+module.exports = {
+  name: 'custom-routes',
+  hook: '$FASTIFY_ROUTE',
+  handler: [
+    {
+      method: 'GET',
+      url: '/custom/route/:name',
+      handler: async (request) => `custom: ${request.params.name}`,
+    },
+  ],
+};