+++ packages/kitchensink/service-fastify-cors/service-fastify-cors.test.js
@@ -1,0 +1,6 @@
+describe('service-fastify-cors', () => {
+  it('Should render a custom route', async () => {
+    const res = await get(`/custom/route/foo`);
+    expect(res).toBe('custom: foo');
+  });
+});