+++ packages/service-fastify-cors/src/hooks.js
@@ -1,0 +1,5 @@
+const SERVICE_NAME = `fastify-cors`;
+
+module.exports = {
+  SERVICE_NAME,
+};