+++ packages/service-fastify-cors/src/index.js
@@ -1,3 +1,16 @@
-export { default } from './lib/index'
-export * from './lib/index'
-export * from './lib/hooks'
+const corsPlugin = require('fastify-cors');
+const hooks = require('./hooks');
+
+const onFastifyHacksBefore = ({ registerPlugin }, { getConfig }) => {
+  const options = getConfig('fastify.cors', {});
+  registerPlugin(corsPlugin, options);
+};
+
+module.exports = ({ registerAction }) => {
+  registerAction({
+    hook: '$FASTIFY_PLUGIN',
+    name: hooks.SERVICE_NAME,
+    trace: __filename,
+    handler: onFastifyHacksBefore,
+  });
+};