+++ packages/service-fastify-cors/src/lib/hooks.js
@@ -1,8 +1,0 @@
-const { SERVICE } = require('@forrestjs/hooks');
-
-const SERVICE_NAME = `${SERVICE} fastify-cors`;
-
-module.exports = {
-  SERVICE_NAME
-};
-